+++ PathOfShare/src/main/java/edu/whu/entity/blogs/visitorBlog.java
@@ -3,7 +3,7 @@ package edu.whu.entity.blogs;
 import edu.whu.entity.Blog;
 
 public class visitorBlog extends Blog {
-    public visitorBlog(long userId, String content,String title){
+    public visitorBlog(Integer userId, String content,String title){
         super(userId,content,title);
         setFromWho("visitor");
     }
