+++ PathOfShare/src/main/java/edu/whu/entity/comments/visitorComment.java
@@ -3,7 +3,7 @@ package edu.whu.entity.comments;
 import edu.whu.entity.Comment;
 
 public class visitorComment extends Comment {
-    public visitorComment(long posterId, long blogId, String content){
+    public visitorComment(Integer posterId, Integer blogId, String content){
         super(posterId, blogId, content);
         setFromWho("visitor");
     }
