+++ PathOfShare/src/main/java/edu/whu/entity/users/CommenUser.java
@@ -23,14 +23,14 @@ public class CommenUser extends User {
     }
 
     @Override
-    public Comment postAndGetComment(String content, long blogId) throws SQLException  {
+    public Comment postAndGetComment(String content, Integer blogId) throws SQLException  {
         Comment comment = new userComment(getId(),blogId,content);
         postComment(comment);
         return comment;
     }
 
     @Override
-    public Message sendAndGetMessage(String content, long receiverId) throws SQLException  {
+    public Message sendAndGetMessage(String content, Integer receiverId) throws SQLException  {
         Message message = new userMessage(getId(),receiverId,content);
         sendMessage(message);
         return message;
